Add tag description endpoint and update Question model
Added a new endpoint in TagModule.cs to get the description of a tag by its ID using IMediator and GetTagDescriptionQuery. Modified the Question class in Question.cs to make Title, Slug, and Content properties optional. Added FindTagDescription method to ITagRepository interface in ITagRepository.cs and implemented it in TagRepository.cs. Updated QuestionRepository.cs to remove inclusion of Description property when querying tags. Added GetTagDescriptionQuery and GetTagDescriptionHandler to handle the request and return the tag description.